+++ src/components/InfoCard/InfoCard.stories.tsx
@@ -16,6 +16,8 @@ export const Default = {
     variant: InfoCardVariantTypes.Outlined,
     color: InfoCardColorTypes.Info,
     fontSize: 35,
+    maxHeight: '100vh',
+    minHeight: '50px',
     testId: 'testId',
     showStatus: false,
     showStatusIcon: true,