+++ src/components/InfoCard/InfoCard.tsx
@@ -18,6 +18,8 @@ export interface InfoCardProps {
   clickFunction?: Function;
   color?: InfoCardColorTypes;
   fontSize?: number;
+  maxHeight?: string;
+  minHeight?: string;
   message: string;
   showStatus?: boolean;
   showStatusIcon?: boolean;
@@ -31,6 +33,8 @@ export function InfoCard({
   clickFunction,
   color = InfoCardColorTypes.Info,
   fontSize = 35,
+  maxHeight = '100vh',
+  minHeight = '50px',
   message,
   showStatus = false,
   showStatusIcon = true,
@@ -77,9 +81,25 @@ export function InfoCard({
             </Box>
           )}
           {message && message.length && (
-            <Typography sx={{ fontSize: { fontSize }, display: 'flex', justifyContent: 'center' }}>
-              {message}
-            </Typography>
+            <Box
+              component="div"
+              sx={{
+                overflow: 'auto',
+                minHeight: { minHeight },
+                maxHeight: { maxHeight },
+                whiteSpace: 'normal',
+              }}
+            >
+              <Typography
+                sx={{
+                  fontSize: { fontSize },
+                  display: 'flex',
+                  justifyContent: 'center',
+                }}
+              >
+                {message}
+              </Typography>
+            </Box>
           )}
         </Stack>
       </Alert>